# [hotfix] Improve page load for project Overview pages; optimize v2 node contributors endpoint