--- a/tests/test_views.py
+++ b/tests/test_views.py
@@ -4145,16 +4145,17 @@ class TestWikiWidgetViews(OsfTestCase):
         self.project2.update_node_wiki(name='home', content='', auth=Auth(self.project.creator))
 
     def test_show_wiki_for_contributors_when_no_wiki_or_content(self):
-        assert_true(_should_show_wiki_widget(self.project, self.project.creator))
-        assert_true(_should_show_wiki_widget(self.project2, self.project.creator))
+        contrib = self.project.contributor_set.get(user=self.project.creator)
+        assert_true(_should_show_wiki_widget(self.project, contrib))
+        assert_true(_should_show_wiki_widget(self.project2, contrib))
 
     def test_show_wiki_is_false_for_read_contributors_when_no_wiki_or_content(self):
-        assert_false(_should_show_wiki_widget(self.project, self.read_only_contrib))
-        assert_false(_should_show_wiki_widget(self.project2, self.read_only_contrib))
+        contrib = self.project.contributor_set.get(user=self.read_only_contrib)
+        assert_false(_should_show_wiki_widget(self.project, contrib))
+        assert_false(_should_show_wiki_widget(self.project2, contrib))
 
     def test_show_wiki_is_false_for_noncontributors_when_no_wiki_or_content(self):
-        assert_false(_should_show_wiki_widget(self.project, self.noncontributor))
-        assert_false(_should_show_wiki_widget(self.project2, self.read_only_contrib))
+        assert_false(_should_show_wiki_widget(self.project, None))
 
 
 class TestProjectCreation(OsfTestCase):