<!-- Before submit your Pull Request, make sure you picked
     the right branch:

     - For hotfixes, select "master" as the target branch
     - For new features, select "develop" as the target branch
     - For release feature fixes, select the relevant release branch (release/X.Y.Z) as the target branch -->

## Purpose

<!-- Describe the purpose of your changes -->
Improve response times for requests made to the project overview page.

## Changes

Optimizations to /v2/nodes/<nid>/contributors/

- Use `.include('user__guids')` to prefetch user guids
- Use sparse fieldssets when getting the list of contribs for @mentions 

Optimizations to project overview page

- `include` the contributors for each component
- Remove a number of redundant addon- and registration-related queries


## Side effects

<!--Any possible side effects? -->

Results from local testing:

Requesting http://localhost:8000/v2/nodes/jtcu9/contributors/?page[size]=50 (SIPS project)

Before: 100 queries, ~3 seconds
After: 10 queries, ~270 ms

Page load for http://localhost:5000/jtcu9/ (SIPS project)

Before: ~115 seconds
After: ~25 seconds

## Ticket

<!-- Link to JIRA ticket, if applicable e.g. https://openscience.atlassian.net/browse/OSF-1234 -->

https://openscience.atlassian.net/browse/OSF-8374
https://openscience.atlassian.net/browse/OSF-8370

as a part of https://openscience.atlassian.net/browse/OSF-8356