@sloria 2017-08-07
Going to add a few more optimizations to this.